% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtrac.R
\name{ab_tr_clust}
\alias{ab_tr_clust}
\title{Cluster antibody trajectories}
\usage{
ab_tr_clust(dist_mat, ks, verbose = TRUE, ...)
}
\arguments{
\item{dist_mat}{A distance matrix.}

\item{ks}{A numeric vector of \code{k}s to iterate across.}

\item{verbose}{A logical specifying if progress should be printed.}

\item{...}{Additional arguments to be passed to \code{cluster::pam()}.}
}
\value{
A tibble of \code{pam} objects.
}
\description{
\code{ab_tr_clust()} takes a distance matrix of antibody trajectories (typically the output from \code{ab_tr_dist()}) and assigns trajectories to \code{k} clusters across given \code{ks} using the PAM (Partitioning Around Medioids) algorithm from \code{cluster::pam()}.
}
\examples{
\dontrun{
ab_tr_clust(dist_mat = dist_mat, ks = 1:20, verbose = TRUE, variant = "faster", nstart = 10)
}
}
