% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtrac.R
\name{ab_tr_dist}
\alias{ab_tr_dist}
\title{Compute distance matrix from antibody trajectories}
\usage{
ab_tr_dist(trajectories)
}
\arguments{
\item{trajectories}{A tibble with columns \code{trajectory}, \code{t}, and \code{fc}. See \code{abtrac::trajectories} for an example.}
}
\value{
A distance matrix.
}
\description{
\code{ab_tr_dist()} computes a distance matrix using the \code{cosine*euclidean} distance metric.
}
\examples{
\dontrun{
ab_tr_dist(trajectories = trajectories)
}
}
