% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtrac-package.R
\docType{package}
\name{abtrac-package}
\alias{abtrac}
\alias{abtrac-package}
\title{abtrac: AntiBody TRAjectory Clustering}
\description{
`abtrac` is a small companion package to our paper Prevalent and persistent new-onset autoantibodies in mild to severe COVID-19[^paper]. The package demonstrates the clustering approach taken for classification of new-onset autoantibody trajectories. For details on the rationale and use case, see the paper.
}
\author{
\strong{Maintainer}: August F. Jernbom \email{august.jernbom@scilifelab.se} (\href{https://orcid.org/0000-0002-7773-1851}{ORCID})

}
\keyword{internal}
