% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtrac.R
\name{cluster_lineplot}
\alias{cluster_lineplot}
\title{Lineplots of clusters}
\usage{
cluster_lineplot(trajectories_clust, alpha = 0.2, ncol = NULL, nrow = NULL)
}
\arguments{
\item{trajectories_clust}{Tibble of trajectories and clusters, typically the output of \code{extract_clusters()}. See \code{abtrac::trajectories} for an example of all required columns except \code{cluster}, which is added by \code{extract_clusters()}.}

\item{alpha}{A number (in 0:1) defining the transparency of lines.}

\item{ncol}{A number defining the number of columns for facetting.}

\item{nrow}{A number defining the number of rows for facetting.}
}
\value{
A \code{gg} object.
}
\description{
Makes a \code{cluster_lineplot()} to visualize the obtained clusters of trajectories.
}
\examples{
\dontrun{
cluster_lineplot(trajectories_clust = trajectories_clust, alpha = 0.2, ncol = 4)
}
}
