% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtrac.R
\name{extract_clusters}
\alias{extract_clusters}
\title{Extract clusters for a dplyr::selected \code{k}}
\usage{
extract_clusters(trajectories, pam_objects, sel_k)
}
\arguments{
\item{trajectories}{The original tibble of trajectories used to compute the distance matrix (typically using \code{ab_tr_dist()}). See \code{abtrac::trajectories} for an example.}

\item{pam_objects}{A tibble of \code{pam} objects with columns \code{k} (numeric) and \code{pam_object} (list of \code{pam} objects). Typically the output of \code{ab_tr_clust()}}

\item{sel_k}{The dplyr::selected cluster number \code{k}, possibly dplyr::selected using the silhouette method (see \code{silhouette_plot()}).}
}
\value{
\code{trajectories} with an added column of \code{cluster} numbers.
}
\description{
\code{extract_clusters()} extracts the cluster assignment from a tibble of \code{pam} objects and adds them to the original tibble of trajectories.
}
\examples{
\dontrun{
extract_clusters(trajectories = trajectories, pam_objects = pam_objects, sel_k = 7)
}
}
