% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abtrac.R
\name{silhouette_plot}
\alias{silhouette_plot}
\title{Silhouette plot}
\usage{
silhouette_plot(pam_objects, nudge_y = 0.015, text_size = 8)
}
\arguments{
\item{pam_objects}{A tibble of \code{pam} objects with columns \code{k} (numeric) and \code{pam_object} (list of \code{pam} objects). Typically the output of \code{ab_tr_clust()}.}

\item{nudge_y}{Nudge cluster number label.}

\item{text_size}{Size of cluster number label.}
}
\value{
A \code{gg} object.
}
\description{
Makes a \code{silhouette_plot()} from a tibble of \code{pam} objects, typically the output of \code{ab_tr_clust()}. The silhouette method can be used to dplyr::select the optimal cluster number.
}
\examples{
\dontrun{
silhouette_plot(pam_objects = pam_objects)
}
}
