% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trajectories}
\alias{trajectories}
\title{Example antibody trajectories}
\format{
\subsection{\code{trajectories}}{

A tibble
}
}
\source{
P. Nilsson lab, SciLifeLab, Stockholm, Sweden
}
\usage{
trajectories
}
\description{
Data generated using a bead array in the Nilsson lab.
}
\keyword{datasets}
